Refaktorering - gjør koden lettere å lese. Gi nytt navn til .env vaiabel: REACT_APP_BRUK_AUTENTISERING_I_DEV. Flytt all autentiseringskode inn i auth.tsx(util).
