--- src/authentication/pingApi.ts
@@ -1,10 +1,0 @@
-import axios from 'axios';
-import Environment from '../Environment';
-
-const pingApi = () => {
-  return axios.get(`${Environment().apiUrl}/api/ping`, {
-    withCredentials: true,
-  });
-};
-
-export default pingApi;
